ev_car/pyproject.toml
[project]
name = "ev-car"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.66.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "missingno>=0.5.2",
    "nest-asyncio>=1.6.0",
    "numpy>=2.4.2",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scrapy>=2.14.1",
    "seaborn>=0.13.2",
    "selenium>=4.41.0",
    "statsmodels>=0.14.6",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
    "xgboost>=3.2.0",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.5.5",
    "nbconvert>=7.16.0",
    "nbformat>=5.10.0",
    "scipy-stubs>=1.17.1.2",
]
